+++ app/src/main/kotlin/io/github/xnovo3000/openweather/model/TemperatureUnit.kt
@@ -6,11 +6,11 @@ import io.github.xnovo3000.openweather.R
 enum class TemperatureUnit(
     @StringRes val stringRes: Int,
     val queryName: String,
-    val converter: (Double) -> Double
+    val converter: (Int) -> Int
 ) {
 
     CELSIUS(R.string.temperature_unit_celsius, "celsius", { it }),
-    FAHRENHEIT(R.string.temperature_unit_fahrenheit, "fahrenheit", { it * 1.8 + 32 });
+    FAHRENHEIT(R.string.temperature_unit_fahrenheit, "fahrenheit", { (it * 1.8 + 32).toInt() });
 
     companion object {
